requirements.txt
torch
transformers
langchain==0.0.162
openai==0.27.7
tiktoken==0.4.0
pinecone-client[grpc]==2.2.2
PyMuPDF
nltk
tqdm
python-dotenv
datasets==2.12.0
apache_beam
mwparserfromhell
google-generativeai
requests 